Move schema related logic out of OpenAPI generator
